workshops/ai-coding-workshop/cline/scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.37.14",
    "pyyaml>=6.0.2",
]
